fix: Format on pre-build to fix publish flow
As it stands, the pre_build script generates code which might not be in
keeping with the output of `black`. Therefore the `publish` flow can
fail as any uncommited local changes on `twine upload` attempt to
generate a version with `+XXXX` appended (representing local changes),
which is not accepted by mypy. This change follows the unasync with a
call to our formatter to ensure no diff will be present.

Signed-off-by: Sam Lock <sam@swlock.co.uk>