enhancement: Multiple actions support for PlanResources (#66)
Adds support for calling PlanResources with multiple actions.

This is a breaking change for http client users because there's no nice
way of supporting both the single and multiple actions variants of the
method. Because we use data classes for request and response objects,
if I want to add both `action` and `actions` as optional parameters with
default values, I'd have to move them to the end of the parameter list
which would be a breaking change anyway. The other option is to add a
new method like `plan_resources_multiple_actions` but that's not quite
so nice either. I am not a Python master by any means so I might have
missed an obvious trick here. I am open to suggestions if there's a 
cleaner, non-breaking way to do this.

Fixes #65 

Signed-off-by: Charith Ellawala <charith@cerbos.dev>

---------

Signed-off-by: Charith Ellawala <charith@cerbos.dev>
Signed-off-by: Sam Lock <sam@swlock.co.uk>
Co-authored-by: Sam Lock <sam@swlock.co.uk>